Revert "Refactor/organization (#2)"
This reverts commit 53cc8101dc6e8b58f416bc5fde440e3b083346c4.
