+++ jax_russell/__init__.py
@@ -4,8 +4,7 @@ __author__ = """Sean Easter"""
 __email__ = 'sean@easter.ai'
 __version__ = '0.1.0'
 
-from jax_russell.base import ValuationModel
-from jax_russell.trees import CRRBinomialTree, ExerciseValuer, MaxValuer, SoftplusValuer
+from jax_russell.trees import CRRBinomialTree, ExerciseValuer, MaxValuer, SoftplusValuer, ValuationModel
 
 __all__ = [
     "CRRBinomialTree",
