--- tests/trees/__init__.py
@@ -1,5 +1,0 @@
-"""Testing specific to tree models."""
-from jax_russell import trees
-
-
-tree_classes = [trees.CRRBinomialTree, trees.RendlemanBartterBinomialTree]
