--- tests/trees/test_calc_time_steps.py
@@ -1,23 +1,0 @@
-"""Test time step calculator."""
-from jax_russell import trees
-
-
-from jax import numpy as jnp
-
-
-def test_calc_time_steps():
-    """Test calc_time_steps."""
-    trees.calc_time_steps(
-        1e-3,
-        trees.CRRBinomialTree,
-        ("american",),
-        (
-            jnp.array([100.5]),
-            jnp.array([0.3]),
-            jnp.array([0.5]),
-            jnp.array([0.0]),
-            jnp.array([95.0]),
-            jnp.array([0.08]),
-            jnp.array([0.08]),
-        ),
-    )
